feat(content): migrate AdminContentView logic into reusable useAdminContent composable
This update extracts the entire business logic, state management, filters,
pagination, CRUD operations, PDF handling, preview logic and utility helpers
from AdminContentView.vue into a new composable: useAdminContent().

Key changes:
- Created `src/composables/useAdminContent.js` encapsulating:
  • documents loading, filtering and sorting
  • pagination logic
  • status toggling and delete confirmation logic
  • create/edit dialog management
  • full PDF viewer logic (zoom, pan, page navigation, error handling)
  • preview modal handling
  • utilities for authors/editors, type formatting, status mapping, URLs, dates

- Simplified `AdminContentView.vue` to a clean UI-only component
  with imports mapping directly to composable state and methods.

- Improved reusability, readability and maintainability by centralizing all
  shared logic into the composable, following Vue 3 best practices.

- Added watchers and lifecycle bindings inside the view to react to store
  changes (path/type) and bind global listeners (resize, keyboard navigation).

The view is now significantly smaller, easier to reason about, and structured
around a single source of truth for document management.
