+++ src/layouts/TestLayout.vue
@@ -1,0 +1,15 @@
+<template>
+<main>
+
+</main>
+
+
+
+</template>
+<script setup>
+import ContentListCardMobile from '@/components/content/ContentListCardMobile.vue';
+import SidebarMenu from '@/components/common/SidebarMenu.vue';
+import AppHeader from '@/components/common/AppHeader.vue';
+
+
+</script>
